# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Workout Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WorkoutPlan {
  id: number;
  name: string;
  /** 0=Sunday, 1=Monday, ..., 6=Saturday */
  dayOfWeek: number;
  notes?: string | null;
  createdAt: string;
}

export type ExerciseCategory =
  (typeof ExerciseCategory)[keyof typeof ExerciseCategory];

export const ExerciseCategory = {
  Chest: "Chest",
  Back: "Back",
  Legs: "Legs",
  Shoulders: "Shoulders",
  Arms: "Arms",
  Core: "Core",
  Cardio: "Cardio",
} as const;

export type ExerciseGroup = (typeof ExerciseGroup)[keyof typeof ExerciseGroup];

export const ExerciseGroup = {
  none: "none",
  superset: "superset",
  circuit: "circuit",
} as const;

export interface Exercise {
  id: number;
  planId: number;
  name: string;
  category: ExerciseCategory;
  sets: number;
  reps: number;
  weight?: number | null;
  restSeconds: number;
  videoUrl?: string | null;
  groupType: ExerciseGroup;
  groupId?: string | null;
  sortOrder: number;
}

export type WorkoutPlanWithExercises = WorkoutPlan & {
  exercises: Exercise[];
};

export interface CreateWorkoutPlan {
  name: string;
  dayOfWeek: number;
  notes?: string | null;
}

export interface CreateExercise {
  planId: number;
  name: string;
  category: ExerciseCategory;
  sets: number;
  reps: number;
  weight?: number | null;
  restSeconds: number;
  videoUrl?: string | null;
  groupType: ExerciseGroup;
  groupId?: string | null;
  sortOrder: number;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  active: "active",
  completed: "completed",
  aborted: "aborted",
} as const;

export interface WorkoutSession {
  id: number;
  planId: number;
  planName: string;
  status: SessionStatus;
  startedAt: string;
  completedAt?: string | null;
  notes?: string | null;
  totalVolume?: number | null;
}

export interface SetLog {
  id: number;
  sessionId: number;
  exerciseId: number;
  exerciseName: string;
  setNumber: number;
  repsCompleted?: number | null;
  weightUsed?: number | null;
  completed: boolean;
  loggedAt: string;
}

export type WorkoutSessionDetail = WorkoutSession & {
  setLogs: SetLog[];
};

export interface ActiveSessionResponse {
  session: WorkoutSessionDetail | null;
}

export interface CreateSession {
  planId: number;
}

export interface UpdateSession {
  status: SessionStatus;
  notes?: string | null;
}

export interface LogSet {
  exerciseId: number;
  setNumber: number;
  repsCompleted?: number | null;
  weightUsed?: number | null;
  completed: boolean;
}

export interface UpdateSetLog {
  repsCompleted?: number | null;
  weightUsed?: number | null;
  completed: boolean;
}

export interface StatsOverview {
  totalWorkouts: number;
  weeklyVolume: number;
  totalVolume: number;
  avgWorkoutsPerWeek: number;
  totalSets: number;
}

export interface StreakData {
  currentStreak: number;
  longestStreak: number;
  lastWorkoutDate?: string | null;
}

export interface HeatmapEntry {
  /** ISO date string (YYYY-MM-DD) */
  date: string;
  count: number;
}

export interface PersonalRecord {
  exerciseId: number;
  exerciseName: string;
  category: ExerciseCategory;
  maxWeight?: number | null;
  maxReps?: number | null;
  maxSets?: number | null;
  achievedAt?: string | null;
}

export interface ExerciseProgressEntry {
  date: string;
  avgWeight?: number | null;
  maxWeight?: number | null;
  avgReps?: number | null;
  totalSets: number;
}

export interface CategoryBreakdown {
  category: ExerciseCategory;
  exerciseCount: number;
  totalSets: number;
  totalVolume: number;
}

export interface WeeklySummary {
  weekStart: string;
  weekEnd: string;
  totalWorkouts: number;
  totalVolume: number;
  totalSets: number;
  exerciseBreakdown: CategoryBreakdown[];
}

export type MonthlySummaryWeeklyBreakdownItem = {
  week: string;
  workouts: number;
  volume: number;
};

export interface MonthlySummary {
  month: string;
  totalWorkouts: number;
  totalVolume: number;
  totalSets: number;
  exerciseBreakdown: CategoryBreakdown[];
  weeklyBreakdown: MonthlySummaryWeeklyBreakdownItem[];
}

export type UserProfileUnit =
  (typeof UserProfileUnit)[keyof typeof UserProfileUnit];

export const UserProfileUnit = {
  kg: "kg",
  lbs: "lbs",
} as const;

export interface UserProfile {
  id: number;
  name: string;
  weight?: number | null;
  height?: number | null;
  fitnessGoal?: string | null;
  unit: UserProfileUnit;
  createdAt: string;
}

export type UpdateUserProfileUnit =
  (typeof UpdateUserProfileUnit)[keyof typeof UpdateUserProfileUnit];

export const UpdateUserProfileUnit = {
  kg: "kg",
  lbs: "lbs",
} as const;

export interface UpdateUserProfile {
  name: string;
  weight?: number | null;
  height?: number | null;
  fitnessGoal?: string | null;
  unit: UpdateUserProfileUnit;
}

export type GetExercisesParams = {
  planId?: number;
};

export type GetSessionsParams = {
  limit?: number;
  offset?: number;
};

export type GetHeatmapParams = {
  year?: number;
};