# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Workout Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workout-plans:
    get:
      operationId: getWorkoutPlans
      tags: [workout-plans]
      summary: Get all workout plans
      responses:
        "200":
          description: List of workout plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkoutPlan"
    post:
      operationId: createWorkoutPlan
      tags: [workout-plans]
      summary: Create a workout plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutPlan"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlan"
  /workout-plans/{id}:
    get:
      operationId: getWorkoutPlan
      tags: [workout-plans]
      summary: Get a workout plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workout plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlanWithExercises"
    put:
      operationId: updateWorkoutPlan
      tags: [workout-plans]
      summary: Update a workout plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkoutPlan"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkoutPlan"
    delete:
      operationId: deleteWorkoutPlan
      tags: [workout-plans]
      summary: Delete a workout plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /exercises:
    get:
      operationId: getExercises
      tags: [exercises]
      summary: Get exercises (optionally filtered by plan)
      parameters:
        - name: planId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
    post:
      operationId: createExercise
      tags: [exercises]
      summary: Create an exercise
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExercise"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
  /exercises/{id}:
    put:
      operationId: updateExercise
      tags: [exercises]
      summary: Update an exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExercise"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
    delete:
      operationId: deleteExercise
      tags: [exercises]
      summary: Delete an exercise
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /stats/heatmap:
    get:
      operationId: getHeatmap
      tags: [stats]
      summary: Get calendar heatmap data for workout consistency
      parameters:
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Heatmap data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HeatmapEntry"
...
  /stats/exercise-progress/{exerciseId}:
    get:
      operationId: getExerciseProgress
      tags: [stats]
      summary: Get progress data for a specific exercise
      parameters:
        - name: exerciseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exercise progress data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExerciseProgressEntry"
...
components:
  schemas:
...
    WorkoutPlan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        dayOfWeek:
          type: integer
          description: "0=Sunday, 1=Monday, ..., 6=Saturday"
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, dayOfWeek, createdAt]
    WorkoutPlanWithExercises:
      allOf:
        - $ref: "#/components/schemas/WorkoutPlan"
        - type: object
          properties:
            exercises:
              type: array
              items:
                $ref: "#/components/schemas/Exercise"
          required: [exercises]
    CreateWorkoutPlan:
      type: object
      properties:
        name:
          type: string
        dayOfWeek:
          type: integer
        notes:
          type: string
          nullable: true
      required: [name, dayOfWeek]
    ExerciseCategory:
      type: string
      enum: [Chest, Back, Legs, Shoulders, Arms, Core, Cardio]
    ExerciseGroup:
      type: string
      enum: [none, superset, circuit]
    Exercise:
      type: object
      properties:
        id:
          type: integer
        planId:
          type: integer
        name:
          type: string
        category:
          $ref: "#/components/schemas/ExerciseCategory"
        sets:
          type: integer
        reps:
          type: integer
        weight:
          type: number
          nullable: true
        restSeconds:
          type: integer
        videoUrl:
          type: string
          nullable: true
        groupType:
          $ref: "#/components/schemas/ExerciseGroup"
        groupId:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required: [id, planId, name, category, sets, reps, restSeconds, groupType, sortOrder]
    CreateExercise:
      type: object
      properties:
        planId:
          type: integer
        name:
          type: string
        category:
          $ref: "#/components/schemas/ExerciseCategory"
        sets:
          type: integer
        reps:
          type: integer
        weight:
          type: number
          nullable: true
        restSeconds:
          type: integer
        videoUrl:
          type: string
          nullable: true
        groupType:
          $ref: "#/components/schemas/ExerciseGroup"
        groupId:
          type: string
          nullable: true
        sortOrder:
          type: integer
      required: [planId, name, category, sets, reps, restSeconds, groupType, sortOrder]
...
    HeatmapEntry:
      type: object
      properties:
        date:
          type: string
          description: ISO date string (YYYY-MM-DD)
        count:
          type: integer
      required: [date, count]
...
    ExerciseProgressEntry:
      type: object
      properties:
        date:
          type: string
        avgWeight:
          type: number
          nullable: true
        maxWeight:
          type: number
          nullable: true
        avgReps:
          type: number
          nullable: true
        totalSets:
          type: integer
      required: [date, totalSets]
...