# lib/api-zod/src/generated/types/exercise.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Workout Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseCategory } from "./exerciseCategory";
import type { ExerciseGroup } from "./exerciseGroup";

export interface Exercise {
  id: number;
  planId: number;
  name: string;
  category: ExerciseCategory;
  sets: number;
  reps: number;
  weight?: number | null;
  restSeconds: number;
  videoUrl?: string | null;
  groupType: ExerciseGroup;
  groupId?: string | null;
  sortOrder: number;
}