# lib/api-zod/src/generated/types/exerciseCategory.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Workout Tracker API
 * OpenAPI spec version: 0.1.0
 */

export type ExerciseCategory =
  (typeof ExerciseCategory)[keyof typeof ExerciseCategory];

export const ExerciseCategory = {
  Chest: "Chest",
  Back: "Back",
  Legs: "Legs",
  Shoulders: "Shoulders",
  Arms: "Arms",
  Core: "Core",
  Cardio: "Cardio",
} as const;