# lib/api-zod/src/generated/types/exerciseGroup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Workout Tracker API
 * OpenAPI spec version: 0.1.0
 */

export type ExerciseGroup = (typeof ExerciseGroup)[keyof typeof ExerciseGroup];

export const ExerciseGroup = {
  none: "none",
  superset: "superset",
  circuit: "circuit",
} as const;