# lib/api-zod/src/generated/types/exerciseProgressEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Workout Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ExerciseProgressEntry {
  date: string;
  avgWeight?: number | null;
  maxWeight?: number | null;
  avgReps?: number | null;
  totalSets: number;
}