# lib/api-zod/src/generated/types/workoutPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Workout Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkoutPlan {
  id: number;
  name: string;
  /** 0=Sunday, 1=Monday, ..., 6=Saturday */
  dayOfWeek: number;
  notes?: string | null;
  createdAt: Date;
}