# lib/api-zod/src/generated/types/workoutPlanWithExercises.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Workout Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";
import type { WorkoutPlan } from "./workoutPlan";

export type WorkoutPlanWithExercises = WorkoutPlan & {
  exercises: Exercise[];
};